requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==22.0.0
python-dotenv==1.0.1
Pillow
redis==5.1.0
django-redis==5.4.0
celery==5.3.6
sentry-sdk==1.40.6
phonenumbers>=8.13.0
trio
httpx
bs4
termcolor
tqdm
